requirements.txt
streamlit
google-cloud-firestore
gcsfs
st-files-connection
firebase
pyrebase4 
firebase-admin